декомпозирует редьюсер
