+++ package.json
@@ -32,11 +32,13 @@
     "node": "10"
   },
   "dependencies": {
+    "axios": "0.19.2",
     "prop-types": "15.7.2",
     "react": "16.12.0",
     "react-dom": "16.12.0",
     "react-redux": "7.2.0",
     "react-router-dom": "5.1.2",
-    "redux": "4.0.5"
+    "redux": "4.0.5",
+    "redux-thunk": "2.3.0"
   }
 }
