+++ src/reducer/data/data.js
@@ -1,0 +1,33 @@
+import {extend} from "../../utils.js";
+
+
+const initialState = {
+  questions: [],
+};
+
+const ActionType = {
+  LOAD_QUESTIONS: `LOAD_QUESTIONS`,
+};
+
+const ActionCreator = {
+  loadQuestions: (questions) => {
+    return {
+      type: ActionType.LOAD_QUESTIONS,
+      payload: questions,
+    };
+  },
+};
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.LOAD_QUESTIONS:
+      return extend(state, {
+        questions: action.payload,
+      });
+  }
+
+  return state;
+};
+
+
+export {reducer, ActionType, ActionCreator};
