+++ src/reducer/reducer.js
@@ -1,81 +1,12 @@
-import {extend} from "../utils";
-import {GameType} from "../const";
-import questions from "../mocks/questions";
+import {combineReducers} from "redux";
+import {reducer as data} from "./data/data.js";
+import {reducer as game} from "./game/game.js";
+import {reducer as user} from "./user/user.js";
+import NameSpace from "./name-space.js";
 
-const initialState = {
-  mistakes: 0,
-  maxMistakes: 3,
-  step: -1,
-  questions,
-};
 
-const ActionType = {
-  INCREMENT_MISTAKES: `INCREMENT_MISTAKES`,
-  INCREMENT_STEP: `INCREMENT_STEP`,
-  RESET: `RESET`,
-};
-
-const isArtistAnswerCorrect = (question, userAnswer) => {
-  return userAnswer.artist === question.song.artist;
-};
-
-const isGenreAnswerCorrect = (question, userAnswer) => {
-  return userAnswer.every((it, i) => {
-    return it === (question.answers[i].genre === question.genre);
-  });
-};
-
-const ActionCreator = {
-  incrementStep: () => ({
-    type: ActionType.INCREMENT_STEP,
-    payload: 1,
-  }),
-
-  incrementMistake: (question, userAnswer) => {
-    let answerIsCorrect = false;
-
-    switch (question.type) {
-      case GameType.ARTIST:
-        answerIsCorrect = isArtistAnswerCorrect(question, userAnswer);
-        break;
-      case GameType.GENRE:
-        answerIsCorrect = isGenreAnswerCorrect(question, userAnswer);
-        break;
-    }
-
-    return {
-      type: ActionType.INCREMENT_MISTAKES,
-      payload: answerIsCorrect ? 0 : 1,
-    };
-  },
-
-  resetGame: () => {
-    return {
-      type: ActionType.RESET,
-      payload: null,
-    };
-  }
-};
-
-const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case ActionType.INCREMENT_STEP:
-      return extend(state, {
-        step: state.step + action.payload,
-      });
-
-    case ActionType.INCREMENT_MISTAKES:
-      return extend(state, {
-        mistakes: state.mistakes + action.payload,
-      });
-
-    case ActionType.RESET:
-      return extend(initialState, {
-        step: 0,
-      });
-  }
-
-  return state;
-};
-
-export {reducer, ActionType, ActionCreator};
+export default combineReducers({
+  [NameSpace.DATA]: data,
+  [NameSpace.GAME]: game,
+  [NameSpace.USER]: user,
+});
